pyproject.toml
[project]
name = "llm-bias"
version = "0.1.0"
description = "Position bias in LLMs: multiple-choice QA with option shuffling"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.28.0",
    "openai>=1.0.0",
    "ollama>=0.4.4",
    "huggingface-hub>=0.20.0",
    "datasets>=2.14.0",
    "transformers>=4.36.0",
    "torch>=2.0.0",
    "accelerate>=0.25.0",
    "pyyaml>=6.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.18.0",
    "seaborn>=0.12.0",
    "tqdm>=4.65.0",
    "psycopg[binary]>=3.1.0",
    "python-dotenv>=1.0.0",
    "bitsandbytes>=0.41.0",
]

[project.scripts]
